Move geometry rep codes to a separate file for model simplicity too.
